Отрефакторила утилитарные функции
